+++ lib/ts/recipe/dashboard/types.ts
@@ -72,6 +72,18 @@ export type RecipeLevelUser = {
         id: string;
         userId: string;
     };
+};
+
+export type RecipeLevelUserWithFirstAndLastName = {
+    recipeId: "emailpassword" | "thirdparty" | "passwordless";
+    timeJoined: number;
+    recipeUserId: string;
+    email?: string;
+    phoneNumber?: string;
+    thirdParty?: {
+        id: string;
+        userId: string;
+    };
     firstName: string;
     lastName: string;
 };
