+++ lib/ts/recipe/emailpassword/emaildelivery/services/backwardCompatibility/index.ts
@@ -14,13 +14,11 @@
  */
 import { TypeEmailPasswordEmailDeliveryInput, RecipeInterface } from "../../../types";
 import { createAndSendCustomEmail as defaultCreateAndSendCustomEmail } from "../../../passwordResetFunctions";
-import { NormalisedAppinfo, User } from "../../../../../types";
+import { NormalisedAppinfo } from "../../../../../types";
 import { EmailDeliveryInterface } from "../../../../../ingredients/emaildelivery/types";
-import { getUser } from "../../../../..";
 
 export default class BackwardCompatibilityService
     implements EmailDeliveryInterface<TypeEmailPasswordEmailDeliveryInput> {
-    private recipeInterfaceImpl: RecipeInterface;
     private isInServerlessEnv: boolean;
     private appInfo: NormalisedAppinfo;
     private resetPasswordUsingTokenFeature: {
@@ -29,15 +27,13 @@ export default class BackwardCompatibilityService
                 id: string;
                 recipeUserId: string;
                 email: string;
-                timeJoined: number;
             },
             passwordResetURLWithToken: string,
             userContext: any
         ) => Promise<void>;
     };
 
-    constructor(recipeInterfaceImpl: RecipeInterface, appInfo: NormalisedAppinfo, isInServerlessEnv: boolean) {
-        this.recipeInterfaceImpl = recipeInterfaceImpl;
+    constructor(_: RecipeInterface, appInfo: NormalisedAppinfo, isInServerlessEnv: boolean) {
         this.isInServerlessEnv = isInServerlessEnv;
         this.appInfo = appInfo;
         {
@@ -48,48 +44,18 @@ export default class BackwardCompatibilityService
     }
 
     sendEmail = async (input: TypeEmailPasswordEmailDeliveryInput & { userContext: any }) => {
-        let user: User | undefined =
-            input.user.recipeUserId !== undefined
-                ? await this.recipeInterfaceImpl.getUserById({
-                      userId: input.user.recipeUserId,
-                      userContext: input.userContext,
-                  })
-                : undefined;
-        if (input.user.recipeUserId === undefined) {
-            user = await getUser(input.user.id);
-        }
-        if (user === undefined) {
-            throw Error("this should never come here");
-        }
-        let recipeLoginMethod = user.loginMethods.find(
-            (u) => u.recipeId === "emailpassword" && u.email === input.user.email
-        );
-        if (recipeLoginMethod === undefined) {
-            throw Error("this should never come here");
-        }
         // we add this here cause the user may have overridden the sendEmail function
         // to change the input email and if we don't do this, the input email
         // will get reset by the getUserById call above.
-        let passwordResetEmailUser: {
-            id: string;
-            recipeUserId: string;
-            email: string;
-            timeJoined: number;
-        } = {
-            id: input.user.id,
-            recipeUserId: recipeLoginMethod.recipeUserId,
-            email: input.user.email,
-            timeJoined: recipeLoginMethod.timeJoined,
-        };
         try {
             if (!this.isInServerlessEnv) {
                 this.resetPasswordUsingTokenFeature
-                    .createAndSendCustomEmail(passwordResetEmailUser, input.passwordResetLink, input.userContext)
+                    .createAndSendCustomEmail(input.user, input.passwordResetLink, input.userContext)
                     .catch((_) => {});
             } else {
                 // see https://github.com/supertokens/supertokens-node/pull/135
                 await this.resetPasswordUsingTokenFeature.createAndSendCustomEmail(
-                    passwordResetEmailUser,
+                    input.user,
                     input.passwordResetLink,
                     input.userContext
                 );
