+++ lib/ts/supertokens.ts
@@ -33,15 +33,6 @@ import STError from "./error";
 import { logDebugMessage } from "./logger";
 import { PostSuperTokensInitCallbacks } from "./postSuperTokensInitCallbacks";
 import AccountLinking from "./recipe/accountlinking/recipe";
-import { RecipeLevelUser } from "./recipe/accountlinking/types";
-import EmailPasswordRecipe from "./recipe/emailpassword/recipe";
-import ThirdPartyRecipe from "./recipe/thirdparty/recipe";
-import PasswordlessRecipe from "./recipe/passwordless/recipe";
-import EmailPassword from "./recipe/emailpassword";
-import ThirdParty from "./recipe/thirdparty";
-import Passwordless from "./recipe/passwordless";
-import ThirdPartyEmailPassword from "./recipe/thirdpartyemailpassword";
-import ThirdPartyPasswordless from "./recipe/thirdpartypasswordless";
 
 export default class SuperTokens {
     private static instance: SuperTokens | undefined;
@@ -404,175 +395,5 @@ export default class SuperTokens {
             }
         }
         throw err;
-    };
-
-    // this is an internal use function, therefore it is prefixed with an `_`
-    _getUserForRecipeId = async (
-        userId: string,
-        recipeId: string
-    ): Promise<{
-        user: RecipeLevelUser | undefined;
-        recipe:
-            | "emailpassword"
-            | "thirdparty"
-            | "passwordless"
-            | "thirdpartyemailpassword"
-            | "thirdpartypasswordless"
-            | undefined;
-    }> => {
-        let user: RecipeLevelUser | undefined;
-        let recipe:
-            | "emailpassword"
-            | "thirdparty"
-            | "passwordless"
-            | "thirdpartyemailpassword"
-            | "thirdpartypasswordless"
-            | undefined;
-
-        if (recipeId === EmailPasswordRecipe.RECIPE_ID) {
-            try {
-                const userResponse = await EmailPassword.getUserById(userId);
-
-                if (userResponse !== undefined) {
-                    let loginMethod = userResponse.loginMethods.find(
-                        (u) => u.recipeId === "emailpassword" && u.recipeUserId === userId
-                    );
-                    if (loginMethod !== undefined) {
-                        user = {
-                            ...loginMethod,
-                            recipeId: "emailpassword",
-                        };
-                        recipe = "emailpassword";
-                    }
-                }
-            } catch (e) {
-                // No - op
-            }
-
-            if (user === undefined) {
-                try {
-                    const userResponse = await ThirdPartyEmailPassword.getUserById(userId);
-
-                    if (userResponse !== undefined) {
-                        if ("loginMethods" in userResponse) {
-                            let loginMethod = userResponse.loginMethods.find(
-                                (u) => u.recipeId === "emailpassword" && u.recipeUserId === userId
-                            );
-                            if (loginMethod !== undefined) {
-                                user = {
-                                    ...loginMethod,
-                                    recipeId: "emailpassword",
-                                };
-                                recipe = "thirdpartyemailpassword";
-                            }
-                        } else {
-                            user = {
-                                ...userResponse,
-                                recipeId: "emailpassword",
-                            };
-                            recipe = "thirdpartyemailpassword";
-                        }
-                    }
-                } catch (e) {
-                    // No - op
-                }
-            }
-        } else if (recipeId === ThirdPartyRecipe.RECIPE_ID) {
-            try {
-                const userResponse = await ThirdParty.getUserById(userId);
-
-                if (userResponse !== undefined) {
-                    user = {
-                        ...userResponse,
-                        recipeId: "thirdparty",
-                    };
-                    recipe = "thirdparty";
-                }
-            } catch (e) {
-                // No - op
-            }
-
-            if (user === undefined) {
-                try {
-                    const userResponse = await ThirdPartyEmailPassword.getUserById(userId);
-
-                    if (userResponse !== undefined) {
-                        if ("loginMethods" in userResponse) {
-                            let loginMethod = userResponse.loginMethods.find(
-                                (u) => u.recipeId === "thirdparty" && u.recipeUserId === userId
-                            );
-                            if (loginMethod !== undefined) {
-                                user = {
-                                    ...loginMethod,
-                                    recipeId: "thirdparty",
-                                };
-                                recipe = "thirdpartyemailpassword";
-                            }
-                        } else {
-                            user = {
-                                ...userResponse,
-                                recipeId: "emailpassword",
-                            };
-                            recipe = "thirdpartyemailpassword";
-                        }
-                    }
-                } catch (e) {
-                    // No - op
-                }
-            }
-
-            if (user === undefined) {
-                try {
-                    const userResponse = await ThirdPartyPasswordless.getUserById(userId);
-
-                    if (userResponse !== undefined) {
-                        user = {
-                            ...userResponse,
-                            recipeId: "thirdparty",
-                        };
-                        recipe = "thirdpartypasswordless";
-                    }
-                } catch (e) {
-                    // No - op
-                }
-            }
-        } else if (recipeId === PasswordlessRecipe.RECIPE_ID) {
-            try {
-                const userResponse = await Passwordless.getUserById({
-                    userId,
-                });
-
-                if (userResponse !== undefined) {
-                    user = {
-                        ...userResponse,
-                        recipeId: "passwordless",
-                    };
-                    recipe = "passwordless";
-                }
-            } catch (e) {
-                // No - op
-            }
-
-            if (user === undefined) {
-                try {
-                    const userResponse = await ThirdPartyPasswordless.getUserById(userId);
-
-                    if (userResponse !== undefined) {
-                        user = {
-                            ...userResponse,
-                            recipeId: "passwordless",
-                        };
-                        recipe = "thirdpartypasswordless";
-                    }
-                } catch (e) {
-                    // No - op
-                }
-            }
-        }
-
-        return {
-            user,
-            recipe,
-        };
     };
 }
